+++ js/generate-posts.js
