+++ js/main.js
@@ -1,3 +1,3 @@
-import {createPosts, SIMILAR_POST_COUNT} from './generate-data.js';
-import {renderPosts} from './render-data.js';
-renderPosts (createPosts(SIMILAR_POST_COUNT));
+import {createPosts, SIMILAR_POST_COUNT} from './generate-posts.js';
+import {renderPosts} from './render-posts.js';
+renderPosts(createPosts(SIMILAR_POST_COUNT));
